Created new module to hold all configurations for PMD, Checkstyle and NOTICE generation.
This will be build when building the WebTester's parent project. It does not reference the WebTester parent because that would create a cyclic dependency.
The module is declared as a dependency for the PMD, Checkstyle and Notice plugins in order to allow acces to the included configuration files via the classpath.